+++ Monofoxe.Playground/Monofoxe.Playground/GameController.cs
@@ -2,7 +2,7 @@
 using Microsoft.Xna.Framework.Graphics;
 using Monofoxe.Engine;
 using Monofoxe.Engine.Drawing;
-using Monofoxe.Engine.ECS;
+using Monofoxe.Engine.EC;
 using Monofoxe.Engine.SceneSystem;
 using Monofoxe.Engine.Cameras;
 using Resources.Sprites;
@@ -66,12 +66,13 @@ namespace Monofoxe.Playground
 
 		public override void Update()
 		{
-
+			base.Update();
 		}
 
 
 		public override void Draw()
 		{
+			base.Draw();
 		}
 
 	}