+++ Monofoxe.Playground/Monofoxe.Playground/GraphicsDemo/SpriteDemo.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
-using Monofoxe.Engine.ECS;
+using Monofoxe.Engine.EC;
 using Monofoxe.Engine.Drawing;
 using Monofoxe.Engine.Utils;
 using Monofoxe.Engine;
@@ -43,6 +43,8 @@ namespace Monofoxe.Playground.GraphicsDemo
 
 		public override void Update()
 		{
+			base.Update();
+
 			// Basic animation code.
 			_animation += TimeKeeper.Global.Time(_animationSpeed);
 			if (_animation > 1)
@@ -53,6 +55,8 @@ namespace Monofoxe.Playground.GraphicsDemo
 
 		public override void Draw()
 		{
+			base.Draw();
+
 			var startingPosition = new Vector2(100, 100);
 			var position = startingPosition;
 			var spacing = 100;