+++ Monofoxe/Monofoxe.Engine/GameMgr.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Reflection;
 using Microsoft.Xna.Framework;
-using Monofoxe.Engine.ECS;
+using Monofoxe.Engine.EC;
 using Monofoxe.Engine.SceneSystem;
 using Monofoxe.Engine.Drawing;
 
@@ -94,8 +94,7 @@ namespace Monofoxe.Engine
 			WindowManager = new WindowMgr(game);
 
 			LoadAssembliesAndTypes(game.GetType().Assembly);
-
-			SystemMgr.InitSystemPool();
+			
 			AssetMgr.Init();
 
 			var defScene = SceneMgr.CreateScene("default");